+++ api/delete_seance.php
@@ -6,11 +6,16 @@ if (!isset($_POST['seanceID'])) {
     $result = ['status' => false, 'mess' => "Ошибка удаления сеанса"];
 } else {
     $seanceID = (int)($_POST['seanceID']);
-    $deleteSeance = $db->delete('seances', ['id' => $seanceID]);
-    if ($deleteSeance) {
-        $result = ['status' => true];
+    $hallID = $db->selectWhere('seances', ['hall_id'], ['id' => $seanceID])[0]['hall_id'];
+    if(canChange($hallID)) {
+        $deleteSeance = $db->delete('seances', ['id' => $seanceID]);
+        if ($deleteSeance) {
+            $result = ['status' => true];
+        } else {
+            $result = ['status' => false, 'mess' => print_r($deleteSeance->errorInfo(), true)];
+        }
     } else {
-        $result = ['status' => false, 'mess' => print_r($deleteSeance->errorInfo(), true)];
+        $result = ['status' => false, 'mess' => 'Для удаления сеанса приостановите продажу билетов'];
     }
 }
